Implement Read; generate get (#137)
This change does two basic things.

First, it implements the new Read RPC method on the resource provider
interface.  This simply calls through to the underlying Terraform
provider's Refresh function, marshaling state to and fro as usual.

Second, it generates a static `get` function on all resource types:

    public static get(name: string, id: Input<ID>, state?: RState): R;

This function takes a name, for purposes of URN creation, a resource
ID to use in the lookup, and an optional RState object for further
lookup filtering.  It uses the Read method to lookup and populate the
complete resource state and returns an instance of the enclosing
resource type R.

Note that RState is not known a priori and in fact just contains all
possible properties the given resource could have.  Many resources can
be loaded just given their ID, but certain other complex resources
cannot; the requirements here are highly resource dependent.