Project unknown lists as lists of unknowns. (#170)
Terraform does not understand that an unknown list is in fact a list.
Instead it expects a list of unknown values. These changes perform this
projection.

I verified that for other complex types--notably maps and
objects--Terraform does accept an unknown value as valid.

This should unblock https://github.com/pulumi/pulumi-cloud/pull/506.