Include `id` property in data source outputs
Terraform data source results will always have an ID, this is now
projected to an attribute on the structs returned by data source get
functions.

For AWS and Azure, this ID is nearly always an alias for some other data
projected in a separate field. However, in the Openstack provider, this
ID is frequently the only way that certain information gets returned
from the data source.

As the ID field isn't exposed by terraform schema we special case it as
an attribute in the generator and provider.

Fixes #188