Update pulumi/pulumi. (#214)
This requires adding a new method to the provider server, Cancel, which
requests that the provider abort any ongoing resource operations. We
currently implement this as a no-op. It's not clear that we can
reasonably do any better given TF's provider API, which has no
first-class affordances for cancellation.

Absorbing the current version of pulumi/pulumi also required updating an
asset test. As of pulumi/pulumi@b9a0907, assets can no longer be empty,
so this test is no longer necessary.