Save TF resource metadata in state. (#244)
These changes add support for storing a Terraform resource's metadata
alongside its attributes in its Pulumi state. The metadata is stored as
a Pulumi property map under the key `__meta`, which should not be used
as a property name by a Terraform resource. Notably, this metadata
includes the `schema_version` property, which is used by the TF schema
code to detect the need for schema migration.

If a resource has no metadata but its schema's `SchemaVersion` is
non-zero, we inject that version into its metadata before interacting
with the TF provider. This ensures that TF always sees a valid schema
version when one is required.

This is necessary to address #226.