Look up Pulumi schema by mangled resource name
getInfoFromPulumiName is used as part of the schema resolution in the
Check phase of tfbridge, and the Pulumi Schema is used to determine
whether there is a value transformer or not (as of e741081).

The Pulumi schema is returned by indexing into `ps`, a
map[string]*tfbridge.SchemaInfo. The keys of the map are in the mangled
Terraform underscore_separated format, as is verifiable with a debugger,
in this case for an aws.iam.Role (or aws_iam_role) resource:

```
0.031s ps=map[string]*tfbridge.SchemaInfo{
           "name": &tfbridge.SchemaInfo{
               Name:      "name",
               Type:      "",
               AltTypes:  nil,
               Transform: tfbridge.Transformer {...},
               Elem:      (*tfbridge.SchemaInfo)(nil),
               Fields:    {},
               Asset:     (*tfbridge.AssetTranslation)(nil),
               Default:   &tfbridge.DefaultInfo{
                   From:    func(*tfbridge.PulumiResource) (interface {}, error) {...},
                   Value:   nil,
                   EnvVars: nil,
               },
               Stable:      (*bool)(nil),
               MaxItemsOne: (*bool)(nil),
           },
           "assume_role_policy": &tfbridge.SchemaInfo{
               Name:        "",
               Type:        "string",
               AltTypes:    {"aws:iam/documents:PolicyDocument"},
               Transform:   tfbridge.Transformer {...},
               Elem:        (*tfbridge.SchemaInfo)(nil),
               Fields:      {},
               Asset:       (*tfbridge.AssetTranslation)(nil),
               Default:     (*tfbridge.DefaultInfo)(nil),
               Stable:      (*bool)(nil),
               MaxItemsOne: (*bool)(nil),
           },
       }
```

However, the getInfoFromPulumiName function indexed into `ps` using a
non-mangled name - yielding a Terraform schema but no Pulumi schema, and
consequently no transformer - giving the errors seen in
pulumi/pulumi-aws#356 where a non-transformed value is passed through
erroneously to the Terraform provider.

This commit fixes the lookup to use `name` instead of `ks`, which will
respect the `rawName` argument.