Call Terraform Importer as part of get operations (#284)
This is a reworking of commit b13b8a1b, which was merged into
`pulumi-terraform` but subsequently reverted because of emergent bugs
when refreshing some types of resources.

Terraform Importers exist to allow resources which cannot (or are not)
read purely by ID to populate the necessary parts of configuration
_from_ their ID. Previously, this functionality was unused in the Pulumi
Terraform Bridge, leading to crashes (e.g. pulumi/pulumi-gcp#55) or
inability to import without duplicating data points which form part of
resource IDs (e.g. pulumi/pulumi-gcp#48).

This commit calls the importer as part of the read operation, provided
all of the following conditions are true:

- an importer is defined in the Terraform schema
- the properties map on the Read request is empty, indicating a get
  operation instead of a refresh operation

After running the importer, we add the values set by the importer to the
property bag passed to Read.

This fixes both classes of bugs - validation of IDs done in the importer
now results in error messages being passed back in preference to panics,
and the following program will now import a GCP resource:

```typescript
import * as gcp from "@pulumi/gcp";

const defaultPool = gcp.container.NodePool.get("default-pool",
     "us-central1-a/test-ci/default-pool");
// Note there is no redundant specification of data here.

export const nodePoolVersion = defaultPool.version;
```

Because an importer is capable of fanning out to multiple resource (for
example, an AWS Security Group imports rules in addition to the group),
it is necessary to match the returned data to the ID of the resource in
order to determine a match.

Right now, we ignore nil states returned from the provider - while they
are a Terraform provider bug if they occur, we have the ability to work
around them by specifying properties directly, and it is not desirable
to disable that ability at this stage by erroring.

There is some test coverage to show that this minimally works - it would
be better to refactor some aspects of the provider to make it transport-
agnostic to allow easier testing in future.