Split documentation H1 at first occurence only
When parsing the header section of a documentation page, we were
splitting by "\n# " to find H1 tags represented in Markdown.
Unfortunately, this co-incides with the comment tag in HCL code inside
code fences, and we do not track code fence contexts.

While a better fix for this would be to parse code fences and ignore
their contents, it should be sufficient to take the first H1 element as
the resource name and leave any subsequent ones in the section alone -
assuming they are comments - by using SplitN with n = 2 instead of
Split.

If no H1 lines appear, we still emit a warning as this is not expected.
A notable example of this class of spurious warning is
aws_vpc_peering_connection_options which has code comments inside fenced
blocks in the description.