Generated examples from Terraform docs
This change leverages our tool, tf2pulumi, in library form,
to convert HCL examples to Pulumi TypeScript where possible.

For repos I've tried this on, we get a good success rate --
almost always >50% -- which gives us hundreds of useful examples.

In the event of failure, we ignore and keep moving on, although
the tool does report the # of failures at the end of a run. In
the future, I would imagine us tracking % success rates to ensure
that our docs don't regress without us noticing.

As part of this, I've removed our dependency on our (now unused)
fork of Terraform, and also adopted the same govendor-override
strategy that we're using elsewhere to get the repo in a buildable
state. No new dependencies are introduced, although we now depend
on a fair bit more of Terraform that was previously shaken away.