Use tf2pulumi in CLI form
This changes the Terraform docs generation to using tf2pulumi in
CLI form, rather than as a library. This is due to the significant
complications in versioning Terraform dependencies that results
when needing to consume tf2pulumi from projects that also statically
link in their own version of Terraform. This is complicated further
by circular dependencies from the Terraform engine back to the
providers, often leading to difficult-to-reconcile version constraints.
We are already struggling to deal with the intersection of Dep,
Govendor, and Go Modules, across many large projects, and we don't
desire to make it any worse. As such, we'll simplify our lives and
just add a build-time binary dependency on tf2pulumi.