Don't Clean Terraform Schema in tfbridge
Cleaning the Terraform schema using CleanTerraformSchema results in properties
which are marked Removed by the schema definition being actually removed. Such
properties can be set by CRUD methods for a Terraform resource, however, so we
can't just remove them.

By not cleaning in tfbridge but still retaining the calls to Clean in tfgen, we
should remove the class of errors seen in pulumi/pulumi-gcp#79. This approach
does mean that items which are Removed but still set are saved in state, but
that is not likely to present an issue.