Parse more examples
We previously only tried to parse examples that were obviously inside
of "examples" fragments in the source Markdown. Unfortunately, this
meant we skipped a whole lot of HCL, since it shows up in other places
like front matter sections, inconsistently labeled H2s and H3s, etc.

This change just attempts to parse it all. If we encounter a failure,
we still keep parsing away, and it's up to the caller of parseExamples
to figure out how to proceed. The heuristic I landed on after much
experimentation is to elide entire example sections, but to just keep
the code sample out of the resulting docs for essential sections; this
ensures that we don't skip entire front matter, argument, or attribute
sections, while not ended up with blank example sections that look
funny. This heuristic isn't perfect, however it seems to fare well in
practice across the AWS and Azure providers, at least.

The result is looking pretty great!

This fixes pulumi/docs#784.