Change name of Python name and opts arguments (#320)
* Change name of Python name and opts arguments

The given names of __name__ and __opts__ are valid Markdown, which
confuses our documentation generator and causes us to generate extremely
poor Markdown-flavored documentation. This PR changes the default name
to "resource_name" and the option bag to "opts".

This is a breaking change, so for backwards compatability this change
appends __name__ and __opts__ arguments to the end of every resource's
constructor that aliases resource_name and opts. If a user uses those
parameters, their code works the same as before, except that they will
receive a deprecation warning if they have enabled warnings. __name__
and __opts__ are not documented (so they don't break the docs generator)
but they will continue to work until they are removed in a future
release.

* Go linters teach me to spell