Allow ResourcePrefix to be different to Name
Some Terraform providers have a different name to the prefix on all of the
resources - for example the Google Beta provider is named `google-beta`, but all
of the resources have the prefix `google`. Previously, this caused documentation
generation to fail - we can resolve that by introducing a separate
ResourcePrefix member on ProviderInfo, and using that if it is set.