Add ability to remove array elements which are empty objects
This PR is potential way to fix the issue described in pulumi/pulumi-azure#2511,
and several other variants of it we have seen in the past. It's unclear that
removing empty array elements is always the right option, so we make it opt-in
for resources where it is known to be effective instead of applying it
generally.