Improve error messages for bridge errors (#372)
This commit includes the resource URN in the error messages returned by bridged
Terraform providers under some unexpected conditions (e.g. returning a nil state
and no error during create), and switches from assertion to an actual error
messages in situations which are not uncommon but not indicative of a bug in the
Terraform bridge.

Fixes #128.
Fixes #371.