Don't use the same name as the module for the doc file
When generate go source, we want to add a package level comment about
the package with some links about the provider itself. We do similar
things in python and node, but there we write README.md files, and
they are picked up by our doc generation logic. For go, however, we
write regular go source and allow go-doc to do the right thing.

However, a set our packages already have source files that are the
same name as the packge (but differ only by case). That means reusing
the package name can either lead to two files that differ only by case
or a single file which may have the wrong content.

To work around this, just name the file `_about.go`. This is a little
ugly but unlikely to conflict with any resource we'd generate into the
package