Generate named nested types with doc comments (#456)
Generate named nested types instead of anonymous inline expansions for
TypeScript. This makes it possible to add doc comments to these nested types,
which will be available in the API docs.

The nested types are emitted under a top-level `types` module in `input.ts` or
`output.ts` files, similar to the scheme used by the Kubernetes provider,
which is done to avoid drowning out the types that users actually want to use in
the main modules.

So, for example, the type of the `website` property of `aws.s3.BucketArgs` is
`aws.types.input.s3.BucketWebsite`.

There is one difference compared to the Kubernetes provider: the
Kubernetes provider puts `Args` types in `k8s.types.input`, but these
have already emitted in the main modules here.

The names are generated based on the resource name and property chain. So the
name of the type used for the `objectLockConfiguration` property on
`aws.s3.BucketArgs` is `aws.types.input.s3.BucketObjectLockConfiguration`, and
its `rule` property is typed as a further nested type named
`aws.types.input.s3.BucketObjectLockConfigurationRule`.

The name of types used in arrays are made singular, so if a property on
`FooArgs` is named `rules` and is an array, its type would be `FooRule[]`.