Set delete-before-replace for manually-named resources. (#465)
Determine if a resource is manually named by checking its list of
top-level properties that was could have been populated by an
autoname, but was instead populated manually. If there are any such
properties, indicate that the resource must be deleted before it is
replaced.

These changes also fix a bug in the logic that determined the legacy
replaces list.

Fixes #383.