Upgrade hashicorp/terraform dependency
The Terraform codebase has moved a package since 0.12.7. Any providers
that have upgraded their dependency on Terraform will now fail due to
this package moving

This is not something we are going to be able to roll out to all
providers. It will need to be rolled out provider by provider or the
providers will not build