Rework tfbridge & tfgen to use the Terraform Plugin SDK
This prevents us having to depend on the main hashicorp/terraform repo,
and allows us to bridge Terraform providers which have been updated to
use the standalone plugin SDK instead of the one shipped with the core
of Terraform.

Also note that we no longer depend on the terraform/configs/hcl2shim
package, and instead have embedded the UUID representing an unknown
variable in our source directly as a constant. A change in this value
upstream would break much of the ecosystem and thus is unlikely - so
minor duplication in this case is better than an additional dependency.