Add support for generating .NET SDKs. (#56)
These SDKs are implemented using generated C# source. The directory
structure is similar to that of Node: each module is emitted into a
single directory that contains a single namespace. The set of all
modules is compiled into a single assembly that comprises the SDK.
The build process generates a NuGet package that can be published and
consumed as is standard practice in the .NET community