Fix TransformJSONDoc when passing null due to regression in CoerceTerraformString
The changes made in #180 introduced a regression. When a Pulumi override
was present and there was an empty string, the CoerceTerraformString func
returned nil, rather than the empty string. This was problematic when
there was a coupled TransformJSONDocument with that. We need to change
the functionality to do 2 things:

1. CoerceTerraformString - don't return nil for any empty string override
except boolean. This allows us to manage the Terraform changes to be strings
for bools that can be unset e.g. AssociatePublicIp

2. We now need to understand when a `nil` value is passed to the
TransformJSONDocument that we need to treat this as there is an empty
string being passed to it

Before:

```
▶ pulumi up
Previewing update (dev):
     Type                 Name        Plan     Info
     pulumi:pulumi:Stack  create-dev
     └─ aws:s3:Bucket     test                 1 error

Diagnostics:
  aws:s3:Bucket (test):
    error: preparing urn:pulumi:dev::create::aws:s3/bucket:Bucket::test's old property state: expected string or JSON map; got <nil>

```

After:

```
▶ pulumi up
Previewing update (dev):
     Type                 Name        Plan
     pulumi:pulumi:Stack  create-dev

Resources:
    2 unchanged

Do you want to perform this update? details
  pulumi:pulumi:Stack: (same)
    [urn=urn:pulumi:dev::create::pulumi:pulumi:Stack::create-dev]

```