Move example conversion into a second pass.
This is in preparation for the fix for #212. The fix for that issue will
require that we produce an intermediate schema that can be used while
converting code samples for inclusion in the final schema. Code samples
in the intermediate schema will remain in their original format. The
intermediate schema will not be persisted.