Add a TF SDK shim.
These changes add a shim for the `terraform-plugin-sdk` schema,
resource, and provider-related types. This shim provides the
functionality necessary for the bridge, SDK generator, and tf2pulumi.

This package should allow future changes to `terraform-plugin-sdk` to be
adopted with less effort, and may be a reasonable starting point for
speaking the TF gRPC API directly.