Add a gRPC-backed shim implementation. (#267)
This implementation of the shim uses the TF plugin gRPC interface to
implement provider operations. The gRPC interface exposes its own schema
system that is independent of the TF plugin SDK and communicates
configuration and state using `msgpack`-encoded `cty` values. The bulk
of the work is in translating the gRPC schema to the shim's schema,
encoding and decoding cty values to and from Go values, and computing
flatmap-style diffs.