Support provider-side preview. (#276)
Read the proposed state for a Create or Update out of the TF diff.

The field reader implementations in the plugin SDK do not properly
handle nested unknown values, instead treating such values as empty.
These changes add a local implementation of the field reader that
correctly leaves nested unknown values as unknown.

A local implementation was chosen instead of an implementation in the
plugin SDK fork so that it could be easily scoped to reading values from
an `*InstanceDiff`, which only happens during previews. Naively changing
the upstream field reader would affect all field reads, including those
that happen during updates. This seems unacceptably risky.