Hypothetical fix for #296
This change propagates the computed flag when traversing lists,
maps, and objects. This is required because the set hash code logic
intentionally skips adding elements that have been computed, but
the current code as written will lose the "computed-ness" along the
way (leading to panics because the code will attempt to then convert
the unknown value sentinel string,
"74D93920-ED26-11E3-AC10-0800200C9A66", to its expected type -- such
as a bool, leading to panics like
"panic: interface conversion: interface {} is string, not bool").

Verified with the repro at
https://github.com/pulumi/pulumi/issues/5770.

I also undid https://github.com/pulumi/pulumi-terraform-bridge/pull/293
because *I think* that was fixing a symptom of this issue, and not
the root cause, but I could be entirely mistaken about that
(hopefully the regression tests added as part of that fix will tell us).