Retain replacements on conflicting diffs
This change is a hypothetical fix for pulumi/pulumi#5637. The problem
there is that if an item is removed from a set that implies replacement,
we will overwrite the replaced-ness because we traverse the old and
new properties where the paths for different properties might conflict.
The "fix" is to retain the stronger of the diff kinds encountered.

Note that it's possible there are edge cases in here that conflict
in other ways, like the prior diff being an ADD and the new one being
a DELETE. That seems nonsensical and if so, maybe we want to assert.