Reimplement input extraction.
The original approach to input extraction pulled _all_ possible inputs
from a state object into an input object. Although this approach will
conservatively detect any possible inputs, the object that it produces
may not be usable:

- the allowable values for a property may be different when that
  property is an input vs. when it is an output
- certain combinations of output property values may be allowed where
  the same combinations of input property values are not
- etc.

This has caused serious usability issues, e.g. https://github.com/pulumi/pulumi/issues/6241.

The conservative approach was originally necessary in order for the Pulumi
engine to detect possible differences between the actual and desired state,
as that detection was done based on input differences only. The engine's
approach has since changed to defer to the provider for diffs, and
conservative input calculation is no longer necessary.

These changes rework the algorithm for a better user experience by
taking advantage of the fact that if old inputs are present, their shape
must already be correct. Instead of copying all possible inputs, the
bridge now copies only inputs that are present in the old inputs. If a
property was present in the old inputs and was populated as a default,
its "defaultness" is preserved iff its value has not changed.

The implementation used during CLI imports has also been changed to only
pull in properties that are pure inputs. This should pull in far fewer
inputs than the old approach, and has a better chance of generating
output that is immediately usable.

Contributes to https://github.com/pulumi/pulumi/issues/6241.