Propagate assets in Read.
Ensure that any property that was an asset in the old state is also an
asset in the new state.

NOTE: I'm pretty sure that this will cause problems when diffing if a
property that was an asset has changes. Should add a test for this and
consider what to do in that case. Might warrant changes in the SDK et.
al. to represent virtual assets.