feat: Correct data source name in Python
This commit corrects the name representation of a Terraform
data source if used in a docstring for Python.

Before this change a data source named `aws_some_data_source` would have
been converted to `aws.someDataSource` for Python, which does not follow
the naming standards for Python. After this change the function would be
correctly called `aws.some_data_source`.