tfbridge: handle unidrectional conflicts
Some Terraform providers specify ConflictsWith in one direction but not
the other (e.g., pulumi/pulumi-snowflake#11). Teach applyDefaults to
handle this situation too to avoid spurious conflict warnings.

Arguably, specifying ConflictsWith in one direction but not the other is
a bug in the provider schema, but at the very least it's a mistake
that's easy to make. Since Terraform seems to handle it just fine, it
seems worth handling in Pulumi too.