[tfbridge] Elide empty optional inputs in Read.
When extracting input properties from state as part of a `Read` for
import, elide any optional properties that are set to either the zero
value for their type or to the default value for the property. This
improves the import experience by eliminating properties that should
not be set (and cause Check to fail when they are set).