Panics converting HCL should be a warning
Panics converting HCL are incorrectly output as an error. They should be
a warning instead because an error in output typically implies a
non-zero exit code. HCL that could not convert is output elsewhere in
the bridge is already a warning, so this change also creates more
consistency with behavior elsewhere in the bridge.