Clarify renamed entities in missing docs message
Renamed and aliased entities are internally given the pseudo-name
x_legacy. This can cause confusion in the missing docs message in the
bridge output because there is no mapping for the provider for the name
x_legacy. When we know an entity is aliased or renamed, we now print
the original name with an indication that the entity has been aliases so
the user can more easily locate the source of the issue.