[tfbridge] Retain the defaults list if present. (#514)
The input extraction code that runs during a refresh has a bug that
causes it to drop the default list entirely if no properties were set
using defaults. Unfortunately, the semantics of a property bag that does
not contain a default list are different from those of a property bag
that contains an empty default list. These changes ensure that the list
is not dropped if it exsits, even if it is empty.

Fixes part of #490.