Ensure any customDiff detailed changes are not removed from the diff
Fixes: #533

There was an extra iteration around the old values and it was removing
them from the diff map. This extra iteration was removing items from the
diff to ensure they were not state updates. We have changed this logic
to now not remove the diffs if they are forcing a new resource

After this change, we get the correct diff coming back from the provider:

```
Outputs:
  ~ cert-id: "319417831845702948372234168596306827381" => output<string>
```

This is based on the repro code in the original issue