[sdk/v2] Fill out RawConfig.
These changes update the sdk/v2 shim to fill out the RawConfig field on
InstanceState and InstanceDiff. TF providers may inspect the value of
this field to determine the differences between an attribute that has
a default value because it is was not specified and an attribute that
has been explicitly set to its default value.

Fixes #548.