Experimental support to use the Terraform parser
This is the start of a new tf2pulumi converter that uses the Terraform
parser and object model to parse and understand the Terraform source
code.

I've added this as an option when using PULUMI_EXPERIMENTAL, but have
extended the eject tests to cover both this new converter and the
existing converter.

The new converter works by invoking terraforms config parser and then
converting their object model directly into hcl blocks that will result
in a valid PCL program (It might be better to convert directly into a
PCL object model and then write the program out from that, and maybe
that's doable with the hcl2.model package, but it wasn't clear to me
how).