make/build: Install inside project directory
`make build` currently runs `go install`
with the commands provided by the project.

This has a side effect of modifying the caller's global environment
as part of the build,
just so the pulumi-terraform-bridge-test-provider command is on $PATH.

This sets `GOBIN` in the Makefile to a subdirectory so that `go install`
does not install anything globally,
and adds the directory to `PATH`
so that `make test` picks up the binary.