Remove Makefile dependence completely
Trying to have Makefile behave consistently on Windows and Unix
is a losing battle.

Eliminate that completely by building the test provider
from inside the test.

Since multiple tests rely on this binary,
this builds the binary once in TestMain and stores the path in a global.
To avoid shadowing, the global has an overly verbose name.