fix: Allow aiven provider to compile without fork
Attempting to use the upstream provider directly fails due to the assert removed
here.

I believe this change is safe to make, despite affecting all providers, as the
assert causes tfgen to panic. No provider was previously allowed to use double
underscores, therefore no backwards incompatible change can be caused by this.

Background:

The Aiven provider has data sources such as:

> - `pg_partman_bgw__dot__interval` (String)
> - `pg_partman_bgw__dot__role` (String)
> - `pg_stat_statements__dot__track` (String)
(Copied from their changelog.)

The upstream provider uses a pair of routines to map `.` to `__dot__` and vice
versa, and the patch in our fork of Aiven modifies that mapping to avoid double
underscores:

https://github.com/pulumi/terraform-provider-aiven/commit/296d9e0d

By allowing the bridge to generate the name, the upstream provider can be used
without requiring the fork. The casing behavior is preserved, and `__dot__` is
cased exactly as `_dot_` was before.