Fixes #720: ignore invalid defaults
When an upstream provider defines a schema with a non-nil Default value and a non-nil ValidateFunc check such that the
DefaultValue does not pass the validations according to ValidateFunc, the bridge will now ignore the DefaultValue and
proceed as if it has been unset or nil. This works around issues in creating azure.storage.ManagementPolicy.