Fix recognizing MaxItems=1 list blocks in Plugin Framework
Based on a failing example from AWS, ensure that listvalidator.SizeAtMost(1) directive is recognized at schema
generation time so that Pulumi can decide correctly whether a property has plural name and a List[T] type or a singular
name and a T type.

This is especially important for backwards-compatibility for providers migrating to the Plugin Framework.