Fix parsing compound provider configuration properties
For compound (object, list) provider properties, Pulumi passes JSON-in-proto encoded values over the provider RPC. This
used to be recognized in Configure calls by parsing the vars parameter. This is problematic:

- the vars parameter to Configure is deprecated in favor of the args parameter
- CheckConfig should use the same deserialization logic as Configure
- CheckConfig does not have access to the vars parameter

Because of these limitations, the PR moves Configure and CheckConfig to use the same serializer logic and use the
recommended parameters.

With this switch filtering configuration values by prefix is no longer necessary in the code as the caller (Pulumi CLI)
takes care of that.

JSON-in-proto encoding is still apparently used for the new args parameters, although their signatures permit the
protocol to use more natural pbstruct encodings. The new serializer continues to support this. Masrshal functionality is
added so that CheckConfig can use this encoding to propagate config values modified by PreConfigureCallback back to
Pulumi CLI in the expected encoding.