Permit skipping flaky example generation (#1023)
Adds an option to skip example renderings to avoid issues with provider builds becoming unstable around examples that have a flaky rendering. While we did not take this approach with the original pulumi/pulumi-aws#2477 issue, but opted to patch the upstream provider, the PR still includes this as an experimental API in case it is a time-saver in the future, especially for providers that do not patch the upstream.