Fix returning secret values out of Plugin Framework based functions (#1011)
Fixes #998

The fix works around what appears to be the underlying problem in Pulumi SDKs at least for Go and YAML: pulumi/pulumi#12710 - before the fix, Invokes that returned secret values caused programs to fail. After the fix, the behavior is similar to SDKv2 bridged providers, secret values in Invoke outputs are replaced with their plain underlying values, which introduces the risk of accidentally storing them un-encrypted in the state but ensures that the programs at least can call the invokes without issue.