Implement MarkSchemaSecrets
It is the responsibility of the bridged providers to ensure that any PropertyValue data returned to the engine is marked
with resource.MakeSecret when it corresponds to a secret property in the schema. Previously this responsibility was
carried by the serialiers, but it can be more convenient to reuse if expressed as a pure PropertyValue based
transformation pass. This is added now as a new function MarkSchemaSecrets.