Refactor handling provider-level Configuration values (#1086)
* Refactor handling provider-level Configuration values

The only intentional user-visible change is this: before the refactor, if any nested schema paths for a provider-level
configuration property were sensitive, bridged providers would mark the entire property as secret, and it would be
stored as a secret value in the state, even if the nested properties in question were not set by the user. After the
refactor, the behavior is more precise. Bridged providers will only mark the providers properties as secret if configued
with actual secrets.

The main aim of the refactor is to enable the reuse of the code for parsing provider level configuration (ConfigEncoding) and marking schema secrets (MarkSchemaSecrets) for Plugin Framework based providers. The necessary functions and types are now exported from the package.

ConfigEncoding no longer carries the secret marking responsibility as it is taken over by MarkSchemaSecrets.

---------

Co-authored-by: Ian Wahbe <ian@wahbe.com>