Handle empty SchemaInfo during naming (#1096)
There was an error in our logic when accounting for pluralizing a field that has a
conflict with a set name. We check confirm that the set `.Name != candidate`, but we
forgot to check that `.Name != ""`.

Adding the `.Name != ""` check causes the newly added test cases to pass, and unblocks
enabling `x.AutoAliasing` for `pulumi-azure`.