Remove nested secrets sent as JSON-in-proto to CheckConfig
As became evident in pulumi/pulumi-docker#640 v3.68.0 version of Pulumi and .NET SDK may be sending nested values marked
as secret using an object wrapper and a secret sentinel "1b47061264138c4ac30d75fd1eb44270" while encoding the result in
JSON. Currently this defeats the assumptions made in the bridge and leads to panics. Bridged providers need to at least
tolerate such input while the underyling issues are being fixed.

The fix here makes config_encoding recognize secrets nested in JSON-in-proto encoding and strip them away. This is part
of the fix for the panic in pulumi/pulumi-docker#640 but further work is needed to ensure these secret values are
encrypted in the state file as the engine does not recognize them.