[x/muxer]: Log instead of failing on divergent CheckConfig results
This change makes muxer less strict as it is now able to tolerate providers that do not agree on CheckConfig responses.
What used to be a hard error is downgraded to logging a message at debug level.

The practical use case driving this is that before pulumi/pulumi-terraform-bridge#821 is implemented, pf vs v3 based
providers may not agree on the secret value markers in the CheckConfig response.