Suggest corrections for misspelled config keys (#1165)
This change improves presentation of error messages in providers for the case of the provider encountering configuration keys it does not understand. Known keys are checked for Levenshtein-similar keys to present a "Did you mean .." suggestion to the user. Also, the change fixes presentation of compound_keys; the errors used to mention them as provider:compound_key but the correct form is actually provider:compoundKey.