Ensure missing required provider property failures surface in Pulumi CLI (#1173)
Working with https://github.com/Authress/pulumi-authress provider examples exposed an issue with how Pulumi CLI reports missing required provider properties; it is not telling the user which property is missing. After the changes in the PR we get it to report as follows:

```
Diagnostics:
  pulumi:providers:authress (default_1_1_42_dirty_github_/api.github.com/Authress/pulumi-authress):
  
    error: pulumi:providers:authress resource 'default_1_1_42_dirty_github_/api.github.com/Authress/pulumi-authress' has
    a problem: Provider is missing a required configuration key, try `pulumi config set authress:customDomain`: Your
    Authress custom domain. [Configure a custom domain for Authress  account](https://authress.io/app/#/settings?focus=domain)
    or use the [provided domain](https://authress.io/app/#/api?route=overview).

```

The provider did report these using pulumirpc.ConfigureErrorMissingKeys_MissingKey special protobuf error annotation, but it appears that Pulumi CLI ignores this information when it comes out of CheckConfig method, and only honors it when it comes out of the Configure method. Consequently, when validations moved from Configure to CheckConfig the bridge regressed in how the missing required keys are surfaced.

The issue was detected on Plugin Framework based provides but affects both PF providers and SDKv2 based providers.

Switching from pulumirpc.ConfigureErrorMissingKeys_MissingKey form to reporting as CheckFailure seems to bring the desired UX back.