Allow state upgraders to modify ID
Before this change, Pulumi would correctly run migrations on the resource state if those are specified in the upstream
provider. However, in the rare situation where the migrations would intentionally modify the ID field, the bridged
provider would ignore this modification. After the change, migrations of the ID field are respected.

Fixes #923.